+++ R/misc.R
@@ -150,12 +150,8 @@ first_not_null <- function(...) {
   out
 }
 
-isFALSE <- function(x) {
-  identical(FALSE, x)
-}
-
 isNA <- function(x) {
-  identical(NA, x)
+  length(x) == 1L && is.na(x)
 }
 
 is_equal <- function(x, y, ...) {
@@ -198,6 +194,14 @@ as_one_character <- function(x, allow_na = FALSE) {
     stop2("Cannot coerce ", s, " to a single character value.")
   }
   x
+}
+
+has_rows <- function(x) {
+  isTRUE(nrow(x) > 0L)
+}
+
+has_cols <- function(x) {
+  isTRUE(ncol(x) > 0L)
 }
 
 expand <- function(..., dots = list(), length = NULL) {
@@ -903,4 +907,8 @@ expect_match2 <- function(object, regexp, ..., all = TRUE) {
     "Run theme_set(theme_default()) to use the default bayesplot theme."
   )
   invisible(NULL)
+}
+
+.onLoad <- function(libname, pkgname) {
+  backports::import(pkgname)
 }