+++ R/validate.R
@@ -49,11 +49,12 @@ parse_bf.brmsformula <- function(formula, family = NULL, autocor = NULL,
                                  mv = FALSE, ...) {
   x <- validate_formula(formula, family = family, autocor = autocor)
   mv <- as_one_logical(mv)
+  rescor <- mv && isTRUE(x$rescor)
   mecor <- isTRUE(x$mecor)
   formula <- x$formula
   family <- x$family
   autocor <- x$autocor
-  y <- nlist(formula, family, autocor, mv, mecor) 
+  y <- nlist(formula, family, autocor, mv, rescor, mecor) 
   class(y) <- "brmsterms"
   
   if (check_response) {
@@ -154,6 +155,12 @@ parse_bf.brmsformula <- function(formula, family = NULL, autocor = NULL,
     }
     x$pfix$sigma <- 0
   }
+  if ("nu" %in% valid_dpars && no_nu(y)) {
+    if ("nu" %in% c(names(x$pforms), names(x$pfix))) {
+      stop2("Cannot predict or fix 'nu' in this model.")
+    }
+    x$pfix$nu <- 0
+  }
   disc_pars <- valid_dpars[dpar_class(valid_dpars) %in% "disc"]
   for (dp in disc_pars) {
     # 'disc' is set to 1 and not estimated by default
@@ -166,15 +173,16 @@ parse_bf.brmsformula <- function(formula, family = NULL, autocor = NULL,
   }
   check_fdpars(y$fdpars)
   # check for illegal use of cs terms
-  if (has_cs(y) && !(is.null(family) || allows_cs(family))) {
-    stop2("Category specific effects are only meaningful for ", 
-          "families 'sratio', 'cratio', and 'acat'.")
+  if (has_cs(y) && !(is.null(family) || allow_cs(family))) {
+    stop2("Category specific effects require families ", 
+          "'sratio', 'cratio', or 'acat'.")
   }
   # parse autocor formula
+  y$time <- parse_time(autocor)
   if (!is.null(y$dpars[["mu"]])) {
     y$dpars$mu$autocor <- autocor
+    y$dpars$mu$time <- y$time
   }
-  y$time <- parse_time(autocor)
   
   # make a formula containing all required variables
   lhsvars <- if (resp_rhs_all) all.vars(y$respform)
@@ -198,11 +206,12 @@ parse_bf.mvbrmsformula <- function(formula, family = NULL, autocor = NULL, ...) 
   x$rescor <- isTRUE(x$rescor)
   x$mecor <- isTRUE(x$mecor)
   out <- structure(list(), class = "mvbrmsterms")
-  out$terms <- lapply(x$forms, parse_bf, mv = TRUE, ...)
-  tmp <- list(rescor = x$rescor, mecor = x$mecor)
+  out$terms <- named_list(names(x$forms))
   for (i in seq_along(out$terms)) {
-    out$terms[[i]][names(tmp)] <- tmp
-  } 
+    x$forms[[i]]$rescor <- x$rescor
+    x$forms[[i]]$mecor <- x$mecor
+    out$terms[[i]] <- parse_bf(x$forms[[i]], mv = TRUE, ...)
+  }
   out$allvars <- allvars_formula(lapply(out$terms, "[[", "allvars"))
   # required to find variables used solely in the response part
   lhs_resp <- function(x) deparse_combine(lhs(x$respform)[[2]])
@@ -306,6 +315,7 @@ parse_ad <- function(formula, family = NULL, check_response = TRUE) {
   if (is.family(family) && any(nzchar(families))) {
     str_formula <- formula2str(formula)
     ad <- get_matches("(?<=\\|)[^~]*(?=~)", str_formula, perl = TRUE)
+    valid_ads <- family_info(family, "ad")
     if (length(ad)) {
       ad_terms <- attr(terms(formula(paste("~", ad))), "term.labels")
       for (a in ad_funs) {
@@ -316,9 +326,7 @@ parse_ad <- function(formula, family = NULL, check_response = TRUE) {
             x[[a]] <- paste0("resp_", x[[a]])
           }
           ad_terms <- ad_terms[-matches]
-          ad_fams <- ad_families(a)
-          valid <- ad_fams[1] == "all" || all(families %in% ad_fams)
-          if (!is.na(x[[a]]) && valid) {
+          if (!is.na(x[[a]]) && a %in% valid_ads) {
             x[[a]] <- str2formula(x[[a]])
           } else {
             stop2("Argument '", a, "' is not supported for ", 
@@ -333,7 +341,7 @@ parse_ad <- function(formula, family = NULL, check_response = TRUE) {
               collapse_comma(ad_terms))
       }
     }
-    if (is_wiener(family) && check_response && !is.formula(x$dec)) {
+    if (check_response && "wiener" %in% families && !is.formula(x$dec)) {
       stop2("Addition argument 'dec' is required for family 'wiener'.")
     }
     if (is.mixfamily(family) && (is.formula(x$cens) || is.formula(x$trunc))) {
@@ -585,6 +593,8 @@ as.brmsterms <- function(x) {
     list(family = paste0(families[1], "_mv"), link = "identity"),
     class = c("brmsfamily", "family")
   )
+  info <- get(paste0(".family_", families[1]))()
+  out$family[names(info)] <- info
   out$sigma_pred <- any(ulapply(x$terms, 
     function(x) "sigma" %in% names(x$dpar) || is.formula(x$adforms$se)
   ))
@@ -678,39 +688,6 @@ check_fdpars <- function(x) {
     }
   }
   invisible(TRUE)
-}
-
-ad_families <- function(x) {
-  # names of valid families for addition arguments
-  switch(x, 
-    weights = "all",
-    se = c("gaussian", "student", "skew_normal", "custom"),
-    trials = c("binomial", "zero_inflated_binomial", "custom"),
-    cat = c("cumulative", "cratio", "sratio", "acat", "custom"), 
-    cens = c(
-      "gaussian", "student", "lognormal", "skew_normal",
-      "inverse.gaussian", "binomial", "poisson", 
-      "geometric", "negbinomial", "exponential", "beta",
-      "weibull", "gamma", "exgaussian", "frechet",
-      "asym_laplace", "gen_extreme_value", "shifted_lognormal",
-      "custom"
-    ),
-    trunc = c(
-      "gaussian", "student", "lognormal", "skew_normal",
-      "binomial", "poisson", "geometric", "negbinomial",
-      "exponential", "weibull", "gamma", "inverse.gaussian",
-      "exgaussian", "frechet", "asym_laplace", "beta",
-      "gen_extreme_value", "shifted_lognormal", "custom"
-    ),
-    mi = c(
-      "gaussian", "student", "lognormal", "skew_normal",
-      "inverse.gaussian", "exponential", "weibull", 
-      "gamma", "exgaussian", "frechet", "beta",
-      "asym_laplace", "gen_extreme_value", "custom"
-    ),
-    dec = c("wiener", "custom"),
-    stop2("Addition argument '", x, "' is not supported.")
-  )
 }
 
 allvars_formula <- function(x) {